Fixed missing BigInt error with eslint

+++ src/App.js
@@ -1,10 +1,10 @@
-import * as React from "react";
-import { useEffect } from "react";
-import logo from "./logo.svg";
-import "./App.css";
-import * as p2pClientLib from "celestium-p2p";
-import * as client from "celestium-client-logic";
-import { io } from "socket.io-client";
+import * as React from 'react'
+import { useEffect } from 'react'
+import logo from './logo.svg'
+import './App.css'
+import * as p2pClientLib from 'celestium-p2p'
+import * as client from 'celestium-client-logic'
+import { io } from 'socket.io-client'
 
 function App() {
   useEffect(() => {
@@ -17,7 +17,7 @@ function App() {
     import("celestium-wasm").then(x => {
       client.walletLib = x;
       console.log(client.initNewClient());
-      console.log(x.generateTestBlockchain(Math.round(Date.now() / 1000)));
+      console.log(x.generateTestBlockchain(BigInt(Math.round(Date.now() / 1000))));
       console.log("Current balance: " + x.getBalance());
     });
   }, [])
